Heavy refactoring, seems to work, needs testing
